Lots of updates

+++ Assets/UnityX/Scripts/Extensions/System/DirectoryX.cs
@@ -6,6 +6,7 @@ using System.IO;
 
 public static class DirectoryX {
 	public static void DeleteAllContents (this DirectoryInfo directoryInfo, bool alsoDeleteFolder = true) {
+        if(!directoryInfo.Exists) return;
 		foreach(System.IO.FileInfo file in directoryInfo.GetFiles()) file.Delete();
     	foreach(System.IO.DirectoryInfo subDirectory in directoryInfo.GetDirectories()) subDirectory.Delete(true);
 		if(alsoDeleteFolder) directoryInfo.Delete(true);
